Disallow leading `-` in positional args
This resolves ambiguities where in `<x>... --foo`, `<x>` would consume
all inputs since they are defined as any character groups seperated by
whitespace. Another solution could be to look ahead if the next argument
is an option and if it matches the next parser, but this is awkward and
implicit. if the user still wants to pass `--foo` as the value of `<x>`,
he can do so by escaping it: `"--foo"`.